fix: move upgrader to a specific folder
